samples/python/pyproject.toml
[project]
name = "a2a-samples"
version = "0.1.0"
description = "Agent2Agent samples"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "jwcrypto>=1.5.6",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "sse-starlette>=2.2.1",
    "starlette>=0.46.1",
    "typing-extensions>=4.12.2",
    "uvicorn>=0.34.0",
    "Appium-Python-Client>=5.0.0",
    "selenium>=4.31.0",
    "aiohttp>=3.11.18",
    "aiocache>=0.12.3",
    "quart>=0.20.0",
    "quart-cors>=0.8.0",
    "a2a-samples",
    "click>=8.1.8",
    "httpx>=0.28.1",
#    "langchain-google-genai>=2.0.10",
    "langchain_openai>=0.1.0",
    "langchain_anthropic>=0.3.12",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "langchain>=0.3.24",
    "setuptools>=75.8.0",
    "alibabacloud-eds-aic20230930>=4.3.2,<5.0.0",
    "alibabacloud-tea-openapi>=0.3.12,<1.0.0",
    "alibabacloud-tea-console>=0.0.1,<1.0.0",
    "alibabacloud-tea-util>=0.3.13,<1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["common", "hosts"]

[tool.uv.workspace]
members = [
    "agents/crewai",
    "agents/google_adk",
    "agents/langgraph",
    "hosts/cli",
    "hosts/multiagent",
    "agents/llama_index_file_chat",
    "agents/semantickernel",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.2",
]
